+++ example/dhis2_dart_utils_example.dart
@@ -1,6 +1,0 @@
-import 'package:dhis2_dart_utils/dhis2_dart_utils.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}