+++ lib/src/utils/period_engine/constants/period_types.dart
@@ -1,0 +1,21 @@
+class PeriodTypes {
+  static const daily = 'DAILY';
+  static const weekly = 'WEEKLY';
+  static const weeklyWednesday = 'WEEKLYWED';
+  static const weeklyThursday = 'WEEKLYTHU';
+  static const weeklySaturday = 'WEEKLYSAT';
+  static const weeklySunday = 'WEEKLYSUN';
+  static const weeksThisYear = 'WEEKS_THIS_YEAR';
+  static const biWeekly = 'BIWEEKLY';
+  static const monthly = 'MONTHLY';
+  static const biMonthly = 'BIMONTHLY';
+  static const quarterly = 'QUARTERLY';
+  static const sixMonthly = 'SIXMONTHLY';
+  static const sixMonthlyApril = 'SIXMONTHLYAPR';
+  static const yearly = 'YEARLY';
+  static const financial = 'FINANCIAL';
+  static const financialYearNovember = 'FYNOV';
+  static const financialYearOctober = 'FYOCT';
+  static const financialYearJuly = 'FYJUL';
+  static const financialYearApril = 'FYAPR';
+}