+++ lib/src/utils/period_engine/models/period_utility.dart
@@ -1,0 +1,11 @@
+import '../constants/fixed_period_types.dart';
+import 'period_type.dart';
+
+class PeriodUtility {
+  static getPeriodTypeById(String id) {
+    List<Map<String, dynamic>> periodTypes = [...fixedPeriodTypes];
+    Map<String, dynamic>? periodType = periodTypes
+        .firstWhere((element) => element['id'] == id, orElse: () => {});
+    return PeriodType(periodType);
+  }
+}