+++ lib/src/utils/shared/constants/operators_constants.dart
@@ -1,0 +1,14 @@
+class OperatorsConstants {
+  static List<String> arithmeticOperators = ['+', '-', '*', '/', '%'];
+  static List<String> logicalOperators = [
+    '>=',
+    '===',
+    '==',
+    '<=',
+    '!=',
+    '&&',
+    '||',
+    '>',
+    '<',
+  ];
+}