+++ lib/src/utils/shared/exceptions/data_object_exception.dart
@@ -1,0 +1,10 @@
+class DataObjectException implements Exception {
+  final String cause;
+
+  DataObjectException(this.cause);
+
+  @override
+  String toString() {
+    return 'PROGRAM INDICATOR EXCEPTION: $cause';
+  }
+}