+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 ## 1.0.0
 
-- Initial version.
+- Support for Program rule engine implementation for form data objects.
+- Support for Program indicator engine implementation for form data objects.
+- Improved the period utility by exporting period constants supported in DHIS2.
+- Improvement on examples and package documentation.
+
+## 1.0.0-alpha-1
+
+### Initial release
+
+- Support core utility for generation of DHIS2 uids.
+- Support for DHIS2 periods, period types and utilities.